Почти доделал обработчик для id из битрикса

+++ Gems.Sales.Notifier/UseCases/NotifyTaggedUsers/NotifyTaggedUsersCommand.cs
@@ -2,5 +2,5 @@
 
 namespace Gems.Sales.Notifier.UseCases.NotifyTaggedUsers
 {
-    public record NotifyTaggedUsersCommand(int[] bitrixUserIds) : IRequest;
+    public record NotifyTaggedUsersCommand(long[] bitrixUserIds) : IRequest;
 }
